pyproject.toml
[tool.poetry]
name = "rag-bench"
version = "0.1.0"
description = "A comprehensive suite for implementing, testing, and benchmarking various Retrieval-Augmented Generation (RAG) architectures across diverse datasets and use cases."
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
huggingface-hub = "^0.26.2"
openai = "^1.54.4"
python-dotenv = "^1.0.1"
datasets = "^3.1.0"


[tool.poetry.group.dev.dependencies]
black = {extras = ["jupyter"], version = "24.10.0"}
isort = "5.13.2"
pytest = "8.3.3"
pre-commit = "4.0.1"
llama-cpp-python = "^0.3.1"
ipykernel = "^6.29.5"
jupyterlab = "^4.3.0"
ebooklib = "^0.18"
pandas = "^2.2.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
